feat : 휴가대상자 일괄저장기능 구현

+++ backend/team6/src/main/java/programmers/team6/domain/vacation/repository/VacationInfoRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import programmers.team6.domain.vacation.entity.VacationInfo;
 import programmers.team6.domain.vacation.rule.vacationgranteligiblities.VacationGrantEligibility;
 
-public interface VacationInfoRepository extends JpaRepository<VacationInfo, Long> {
+public interface VacationInfoRepository extends JpaRepository<VacationInfo, Integer> {
 
 	@Query(value = "SELECT COUNT(DISTINCT vi.memberId) FROM VacationInfo vi")
 	long countAllMemberIds();
